backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
PyJWT
psycopg2-binary
